requirements.txt
PyNaCl>=1.5.0
pytest>=7.4.0
pytest-cov>=4.1.0
python-dotenv>=1.0.0
bcrypt>=4.0.1
